23w03a

+++ net/minecraft/network/protocol/game/ClientboundSectionBlocksUpdatePacket.java
@@ -58,7 +58,7 @@ implements Packet<ClientGamePacketListener> {
         friendlyByteBuf.writeBoolean(this.suppressLightUpdates);
         friendlyByteBuf.writeVarInt(this.positions.length);
         for (int i = 0; i < this.positions.length; ++i) {
-            friendlyByteBuf.writeVarLong(Block.getId(this.states[i]) << 12 | this.positions[i]);
+            friendlyByteBuf.writeVarLong((long)Block.getId(this.states[i]) << 12 | (long)this.positions[i]);
         }
     }
